pyproject.toml
[project]
name = "guarantee-email-agent"
version = "0.1.0"
description = "Instruction-driven AI agent for warranty email automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "typer[all]>=0.9.0",
]

[project.scripts]
agent = "guarantee_email_agent.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.30.0",
]
